Add month limit (#117)
* Add Export and Import dialogs

* Add export and import business logic

* Add month limit UI

* Add month limit UI

* Fix bugs with dialog insets

+++ common/src/jvmMain/kotlin/com/finance_tracker/finance_tracker/core/ui/dialogs/DeleteDialog.kt
@@ -1,22 +1,15 @@
 package com.finance_tracker.finance_tracker.core.ui.dialogs
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.Divider
-import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.unit.dp
 import com.finance_tracker.finance_tracker.MR
 import com.finance_tracker.finance_tracker.core.common.imePadding
@@ -31,7 +24,7 @@ internal fun DeleteAlertDialog(
     onCancelClick: () -> Unit,
     onDeleteClick: () -> Unit,
 ) {
-    DialogSurface {
+    AlertDialogSurface {
         DeleteDialog(
             titleEntity = titleEntity,
             onCancelClick = onCancelClick,
@@ -46,7 +39,7 @@ internal fun DeleteBottomDialog(
     onCancelClick: () -> Unit,
     onDeleteClick: () -> Unit,
 ) {
-    DialogSurface {
+    BottomSheetDialogSurface {
         DeleteDialog(
             titleEntity = titleEntity,
             onCancelClick = onCancelClick,
@@ -55,34 +48,6 @@ internal fun DeleteBottomDialog(
                 .navigationBarsPadding()
                 .imePadding()
         )
-    }
-}
-
-@Composable
-fun DialogSurface(
-    withDragBar: Boolean = true,
-    content: @Composable () -> Unit
-) {
-    CoinTheme {
-        Surface(
-            color = CoinTheme.color.backgroundSurface
-        ) {
-            Column(
-                modifier = Modifier.navigationBarsPadding(),
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                if (withDragBar) {
-                    Divider(
-                        modifier = Modifier
-                            .padding(top = 6.dp, bottom = 8.dp)
-                            .size(width = 16.dp, height = 2.dp)
-                            .clip(CircleShape)
-                            .background(CoinTheme.color.content)
-                    )
-                }
-                content()
-            }
-        }
     }
 }
